requirements.txt
Jinja2
python-keystoneclient
python-novaclient
python-openstackclient-base
